+++ x/documents/client/cli/txDocument_test.go
@@ -90,7 +90,7 @@ func TestCmdShareDocument(t *testing.T) {
 			} else {
 				require.NoError(t, err)
 				var resp sdk.TxResponse
-				require.NoError(t, ctx.JSONMarshaler.UnmarshalJSON(out.Bytes(), &resp))
+				require.NoError(t, ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &resp))
 				require.Equal(t, tt.code, resp.Code)
 			}
 		})
@@ -153,7 +153,7 @@ func TestCmdSendDocumentReceipt(t *testing.T) {
 			} else {
 				require.NoError(t, err)
 				var resp sdk.TxResponse
-				require.NoError(t, ctx.JSONMarshaler.UnmarshalJSON(out.Bytes(), &resp))
+				require.NoError(t, ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &resp))
 				require.Equal(t, tt.code, resp.Code)
 			}
 		})